Move about dialog creation and window caption into GtkBrowsingTool
Move the about dialog creation into the base class of all tools,
add #aboutTitle, #windowTitle selectors to GtkBrowsingTool and
make sure it is implemented by the subclasses.

Now the close button is working for all about dialogs, the help
menu entry is refering to the tool name and the window title is
the proper one.